chore: follow up
A few items were missed
Formatting
