+++ packages/app-degree-pages/src/components/DetailPage/components/ProgramDescription/index.js
@@ -1,6 +1,7 @@
-import { sanitizeDangerousMarkup } from "@asu/components-core";
 import PropTypes from "prop-types";
 import React from "react";
+
+import { sanitizeDangerousMarkup } from "../../../../../../../internal";
 
 /** @typedef {import('../../../../core/types/detail-page-types').ProgramDescriptionProps} ProgramDescriptionProps */
