+++ packages/app-degree-pages/src/components/ListingPage/components/ProgramList/ListView/index.js
@@ -1,9 +1,11 @@
 /* eslint-disable react/no-danger, jsx-a11y/no-noninteractive-element-to-interactive-role  */
 // @ts-check
-import { sanitizeDangerousMarkup } from "@asu/components-core";
 import React, { Fragment, useContext, createRef } from "react";
 
-import { idGenerator } from "../../../../../../../../internal";
+import {
+  idGenerator,
+  sanitizeDangerousMarkup,
+} from "../../../../../../../../internal";
 import { AppContext } from "../../../../../core/context";
 import { GRID_PROGRAMS_ID } from "../../../../../core/models";
 import { degreeDataPropResolverService } from "../../../../../core/services";
